server/requirements.txt
asgiref==3.7.2
contourpy==1.1.0
cycler==0.11.0
dill==0.3.7
Django==4.2.4
djangorestframework==3.14.0
fonttools==4.42.0
kiwisolver==1.4.4
matplotlib==3.7.2
mpmath==1.3.0
numpy==1.25.2
packaging==23.1
pbr==5.11.1
Pillow==10.0.0
ply==3.11
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
qiskit==0.44.1
qiskit-aer==0.12.2
qiskit-terra==0.25.1
rustworkx==0.13.1
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
stevedore==5.1.0
symengine==0.9.2
sympy==1.12
typing_extensions==4.7.1
